monitor/requirements.txt
redis
google-cloud-monitoring